requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.0
drf-access-policy==1.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
Faker==37.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
XlsxWriter==3.2.3
